Moved admin team page components folder
